requirements.txt
future==0.16.0
gevent==1.2a1
greenlet==0.4.10
msgpack-python==0.4.8
paramiko==2.0.2
pyzmq==16.0.0
simplejson==3.8.1
zerorpc==0.6.0
requests==2.11.1
PyYAML==3.12
pymongo==3.3.1
Flask-Cors==3.0.2
python-dateutil==2.6.0
